Several changes related to: * modelID as instance variable for all top-level entities * instance id must be derived from hashcode, thus, is never assigned * clone and hashcode method muse use all fields *except* id

+++ src/gov/usgswim/sparrow/test/parsers/PredictionContextTest.java
@@ -100,7 +100,7 @@ public class PredictionContextTest extends TestCase {
 		reader.next();
 		pCon.parse(reader);
 
-		assertEquals(Integer.valueOf(22), pCon.getModelID());
+		assertEquals(Long.valueOf(22), pCon.getModelID());
 		assertEquals("HUC8", pCon.getAnalysis().getGroupBy());
 		assertEquals(3, pCon.getTerminalReaches().getReachIDs().size());
